refactor(protocol): purge dead vocabulary, split storage by domain, typed test assertions
- delete worker-driver event vocab, CronJob.Events, five dead ingress.* event
  descriptors, Command.Result/Command.Events (consumer-swept against main)
- split storage/index.ts into semantic domain files (orchestration, decision
  ledger, perimeter, scheduling, actor infrastructure) behind an unchanged
  barrel; public Storage surface proven identical
- strengthen ~84 broad toThrow() protocol test assertions to typed ZodError,
  delete 9 echo tests that could not fail
- annotate ledger stream registry entries SHIPPED/DORMANT; command stream
  status corrected to dormant (no production publisher)
- schema snapshot + AGENTS.md synced (shrinkage only)
