+++ packages/protocol/src/storage/namespace.ts
@@ -1,0 +1,2 @@
+// This file owns the shared type-only namespace contract for storage domains.
+export namespace Storage {}
