+++ packages/protocol/test/cron/schema.test.ts
@@ -57,53 +57,3 @@ describe("CronJob schemas", () => {
     ).toBe(false);
   });
 });
-
-describe("CronJob events", () => {
-  test("Scheduled schema accepts valid data", () => {
-    expect(
-      CronJob.Events.CronJobScheduled.schema.parse({
-        traceId: "trace-1",
-        time: 1000,
-        jobId: "cron-1",
-        agentName: "scheduler",
-        schedule: "* * * * *",
-      }),
-    ).toEqual({
-      traceId: "trace-1",
-      time: 1000,
-      jobId: "cron-1",
-      agentName: "scheduler",
-      schedule: "* * * * *",
-    });
-  });
-
-  test("Fired schema accepts valid data", () => {
-    expect(
-      CronJob.Events.CronJobFired.schema.parse({
-        traceId: "trace-1",
-        time: 1000,
-        jobId: "cron-1",
-        agentName: "scheduler",
-      }),
-    ).toEqual({
-      traceId: "trace-1",
-      time: 1000,
-      jobId: "cron-1",
-      agentName: "scheduler",
-    });
-  });
-
-  test("Cancelled schema accepts valid data", () => {
-    expect(
-      CronJob.Events.CronJobCancelled.schema.parse({
-        traceId: "trace-1",
-        time: 1000,
-        jobId: "cron-1",
-      }),
-    ).toEqual({
-      traceId: "trace-1",
-      time: 1000,
-      jobId: "cron-1",
-    });
-  });
-});
